% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chap.R
\name{PartosPoisson}
\alias{PartosPoisson}
\title{PartosPoisson}
\usage{
PartosPoisson(x, tiempo, probabilidad = 1)
}
\arguments{
\item{x}{Un vector numérico con los datos de eventos observados.}

\item{tiempo}{El tiempo en unidades de tiempo en el que se desea calcular la probabilidad.}

\item{probabilidad}{El número mínimo de eventos deseados.}
}
\value{
Un mensaje que indica la probabilidad de tener al menos la cantidad especificada de eventos en el tiempo dado.
}
\description{
Calcula la probabilidad de tener al menos cierta cantidad de eventos en un tiempo determinado utilizando la distribución de Poisson.
}
\details{
La función toma como entrada un vector numérico \code{x} que contiene la cantidad de eventos observados en diferentes momentos.
Utiliza estos datos para calcular un valor estimado del parámetro de tasa \code{lambda} utilizando la fórmula \code{lambda <- sum_valores / cantidad_valores},
donde \code{sum_valores} es la suma de los valores de \code{x} y \code{cantidad_valores} es la cantidad de valores no faltantes.
Luego, verifica que \code{x} sea un vector numérico de enteros no negativos y que \code{lambda} sea un número positivo.
A continuación, calcula la probabilidad de tener al menos \code{probabilidad} eventos en \code{tiempo} unidades de tiempo utilizando la fórmula de la distribución de Poisson.
La probabilidad se presenta en forma de porcentaje.

La función también genera una gráfica que muestra la probabilidad en función del tiempo.
El eje x representa el tiempo y el eje y representa la probabilidad.
}
\examples{
# Datos de partos gemelares
partos_gemelares <- c(1, 1, 0, 1, 1, 1, 1, 1, 1, 2)

# Calcular la probabilidad de al menos 2 partos gemelares en 5 unidades de tiempo
resultado <- PartosPoisson(partos_gemelares, tiempo = 5, probabilidad = 2)
print(resultado)

}
